+++ src/app/(landing)/events_light.tsx
